Move `CustomJSONEncoder` out of `ray_utils`
This should make it easier to cleanly import only `CustomJSONEncoder`, without
importing `ray`.
